// frontend/src/core/api/generated/model/siteAttendanceOverview.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Kızılkan Şantiye API
 * OpenAPI spec version: 1.0
 */
import type { AttendanceCounts } from './attendanceCounts'

export interface SiteAttendanceOverview {
  siteId: string
  siteName: string
  workerCount: number
  /** @nullable */
  lastDay?: string | null
  today?: AttendanceCounts
}